Add register page

+++ frontend/src/pages/RegisterPage.jsx
@@ -12,35 +12,53 @@ import { useNavigate } from "react-router-dom";
 import apiClient from "../services/apiClient";
 import { useAuth } from "../context/AuthContext";
 
-export default function LoginPage() {
-  const [email, setEmail] = useState("");
+export default function RegisterPage() {
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
+  const [firstName, setFirstName] = useState("");
+  const [lastName, setLastName] = useState("");
+  const [email, setEmail] = useState("");
   const [error, setError] = useState(null);
+  const [isSubmitting, setIsSubmitting] = useState(false);
 
   const navigate = useNavigate();
   const { login } = useAuth();
 
-  const handleLogin = async (e) => {
+  const handleRegister = async (e) => {
     e.preventDefault();
     setError(null);
 
+    if (password !== confirmPassword) {
+      setError("Passwords do not match");
+      return;
+    }
+
+    setIsSubmitting(true);
     try {
-      const response = await apiClient.post("/accounts/login/", {
+      await apiClient.post("/accounts/register/customer/", {
+        password,
+        email,
+        first_name: firstName,
+        last_name: lastName,
+      });
+
+      const loginResponse = await apiClient.post("/accounts/login/", {
         email,
         password,
       });
-      const { access, refresh } = response.data.tokens;
-      const user = response.data.user;
+      const { access, refresh } = loginResponse.data.tokens;
+      const user = loginResponse.data.user;
 
       localStorage.setItem("refreshToken", refresh);
       login(user, access);
-
       navigate("/");
     } catch (err) {
       const errorMessage =
         err.response?.data?.error?.message ||
-        "Error login. Verify your credentials and try again";
+        "Error register. Verify your credentials and try again";
       setError(errorMessage);
+    } finally {
+      setIsSubmitting(false);
     }
   };
 
@@ -55,9 +73,9 @@ export default function LoginPage() {
         }}
       >
         <Typography component="h1" variant="h5">
-          Login
+          Register
         </Typography>
-        <Box component="form" onSubmit={handleLogin} sx={{ mt: 1 }}>
+        <Box component="form" onSubmit={handleRegister} sx={{ mt: 1 }}>
           {error && (
             <Alert severity="error" sx={{ width: "100%", mb: 2 }}>
               {error}
@@ -76,18 +94,44 @@ export default function LoginPage() {
             margin="normal"
             required
             fullWidth
+            label="First Name"
+            value={firstName}
+            onChange={(e) => setFirstName(e.target.value)}
+          />
+          <TextField
+            margin="normal"
+            required
+            fullWidth
+            label="Last Name"
+            value={lastName}
+            onChange={(e) => setLastName(e.target.value)}
+          />
+          <TextField
+            margin="normal"
+            required
+            fullWidth
             label="Password"
             type="password"
             value={password}
             onChange={(e) => setPassword(e.target.value)}
+          />
+          <TextField
+            margin="normal"
+            required
+            fullWidth
+            label="Confirm Password"
+            type="password"
+            value={confirmPassword}
+            onChange={(e) => setConfirmPassword(e.target.value)}
           />
           <Button
             type="submit"
             fullWidth
             variant="contained"
             sx={{ mt: 3, mb: 2 }}
+            disabled={isSubmitting}
           >
-            Login
+            Register
           </Button>
         </Box>
       </Box>
